[CORNERSTONE] Fixed browser attachments, added preliminary httplib support.
There seemed to have problems when uploading files with the Win32 binary version
of PyCurl, and as it is too much trouble to compile Python and PyCurl myself, I
started to add a minimal httplib client support to support posting with httplib
instead of PyCurl.

As a result, I will split the current "curl" module into a "client" module that
contains the abstract code to manage a client and the httplib implementation of
the client, and "curlclient" for the curl-based client. It should be possible to
choose for each request which client can be used (this will be repercuted to the
browse module).